requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blobfile==3.0.0
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.1
cycler==0.11.0
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.10.0
gast==0.5.3
geoopt==0.5.0
gmpy2==2.1.2
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.12.1
hf-xet==1.1.3
huggingface-hub==0.33.0
idna==3.7
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keras==3.4.1
kiwisolver==1.4.4
kornia==0.7.4
kornia_rs==0.1.7
lxml==5.3.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.0
mistune==3.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml_dtypes==0.4.0
mpi4py==3.1.3
mpmath==1.3.0
namex==0.0.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
numpy==1.26.4
opacus==1.5.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
pipreqs==0.5.0
platformdirs==4.3.8
ply==3.11
prompt_toolkit==3.0.51
protobuf==4.25.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.21.0
Pygments==2.15.1
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
safetensors==0.5.3
scipy==1.14.1
setuptools==69.5.1
sip==6.7.12
six==1.16.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.12
tensorboard==2.17.0
tensorboard_data_server==0.7.0
tensorflow==2.17.0
tensorflow_estimator==2.15.0
termcolor==2.1.0
timm==1.0.15
tinycss2==1.4.0
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.11.0
unicodedata2==15.1.0
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.43.0
wrapt==1.14.1
yarg==0.1.9
